backend/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-framework==0.1.0
django-rest-knox==3.1.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
httpie==0.9.9
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
